% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP.weights.r
\name{SNP.weights}
\alias{SNP.weights}
\title{Compute weights}
\usage{
SNP.weights(
  x,
  hotspots,
  hotspots.intensity = 10,
  lambda = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{x}{a bed matrix}

\item{hotspots}{data frame of hotspots positions}

\item{hotspots.intensity}{min intensity of hotspot}

\item{lambda}{regularisation parameter}

\item{verbose}{verbosity}
}
\value{
A SNP map with weights...
}
\description{
Compute weights
}
\details{
Computes weights on segments separated by hotspots.
}
\examples{
# loading data (503 european individuals from 1KG project x 536k SNPs)
KG <- read.bed.matrix(system.file("extdata", "1KG_genos.bed", package="migou"))
# computing weights
map <- SNP.weights(KG, hotspots = hotspots.hg19)
# computing the weighted GRM
K <- weighted.GRM(KG, map$weight)

}
