% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.GRM.r
\name{weighted.GRM}
\alias{weighted.GRM}
\title{Weighted GRM}
\usage{
weighted.GRM(x, weights = rep(1, ncol(x)), which.snps, autosome.only = TRUE)
}
\arguments{
\item{x}{a bed matrix}

\item{weights}{a vector of SNP weights}

\item{which.snps}{a logical vector}

\item{autosome.only}{TRUE to keep only autosomes}
}
\value{
A weighted GRM
}
\description{
Weighted GRM
}
\examples{
# see example in SNP.weights
}
